+++ src/pat/modal/modal.js
@@ -62,7 +62,10 @@ define([
 
             // Restore focus in case the active element was a child of $el and
             // the focus was lost during the wrapping.
-            document.activeElement.focus();
+            // Only if we have an activeElement, as IE10/11 can have undefined as activeElement
+            if (document.activeElement) {
+                document.activeElement.focus();
+            }
             this._init_handlers();
             this.resize();
             this.setPosition();