+++ changes.md
@@ -4,6 +4,7 @@
 
 - A fix for pat-scroll to scroll up to current scroll container instead of body. 
 - A fix for pat-scroll to await loading of all images before determining the amount to scroll up.
+- A fix for IE10/11 where the modal wouldn`t close anymore due to activeElement being undefined
 
 ## 2.0.13 - Apr. 27, 2016